+++ backend/api/quivr_api/modules/sync/repository/sync_interfaces.py
@@ -2,15 +2,14 @@ from abc import ABC, abstractmethod
 from typing import Any, List, Literal
 from uuid import UUID
 
-from quivr_api.modules.sync.dto.inputs import (
-    SyncFileInput,
-    SyncFileUpdateInput,
-    SyncsActiveInput,
-    SyncsActiveUpdateInput,
-    SyncsUserInput,
-    SyncUserUpdateInput,
-)
-from quivr_api.modules.sync.entity.sync import SyncFile, SyncsActive, SyncsFiles
+from quivr_api.modules.sync.dto.inputs import (SyncFileInput,
+                                               SyncFileUpdateInput,
+                                               SyncsActiveInput,
+                                               SyncsActiveUpdateInput,
+                                               SyncsUserInput,
+                                               SyncUserUpdateInput)
+from quivr_api.modules.sync.entity.sync import (SyncFile, SyncsActive,
+                                                SyncsFiles)
 
 
 class SyncUserInterface(ABC):
@@ -70,7 +69,7 @@ class SyncInterface(ABC):
 
     @abstractmethod
     def update_sync_active(
-        self, sync_id: UUID | str, sync_active_input: SyncsActiveUpdateInput
+        self, sync_id: UUID | int, sync_active_input: SyncsActiveUpdateInput
     ):
         pass